+++ src/common/MlChart/HistogramChart/histogramChart.scss
